Correção Exercicio 4

+++ variaveisoperadores/src/variaveisoperadores/CalculoDiferencaEx4.java
@@ -7,19 +7,19 @@ public class CalculoDiferencaEx4 {
 	public static void main(String[] args) {
 		
 		Scanner leia = new Scanner(System.in);
-		int n1, n2, n3, n4, diferenca;
+		float n1, n2, n3, n4, diferenca;
 		
 		System.out.println("numero1: ");
-		n1 = leia.nextInt();
+		n1 = leia.nextFloat();
 		
 		System.out.println("numero2: ");
-		n2 = leia.nextInt();
+		n2 = leia.nextFloat();
 		
 		System.out.println("numero3: ");
-		n3 = leia.nextInt();
+		n3 = leia.nextFloat();
 		
 		System.out.println("numero4: ");
-		n4 = leia.nextInt();
+		n4 = leia.nextFloat();
 		
 		diferenca = (n1 * n2) - (n3 * n4);
